Clase 19

+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/dto/salida/paciente/DomicilioSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.clinicaodontologica.dto.salida.paciente;
 
 public class DomicilioSalidaDto {
 
-    private Long id;
+    private int id;
     private String calle;
     private int numero;
     private String localidad;
@@ -12,7 +12,7 @@ public class DomicilioSalidaDto {
     public DomicilioSalidaDto() {
     }
 
-    public DomicilioSalidaDto(Long id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(int id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -20,11 +20,11 @@ public class DomicilioSalidaDto {
         this.provincia = provincia;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
